fix: remove config top domain and merge into system
