+++ cmd/rke2.go
@@ -7,7 +7,6 @@ package cmd
 
 import (
 	agentcmd "github.com/michielvha/edgectl/cmd/rke2/agent"
-	configcmd "github.com/michielvha/edgectl/cmd/rke2/config"
 	lbcmd "github.com/michielvha/edgectl/cmd/rke2/lb"
 	servercmd "github.com/michielvha/edgectl/cmd/rke2/server"
 	systemcmd "github.com/michielvha/edgectl/cmd/rke2/system"
@@ -21,9 +20,11 @@ var rke2Cmd = &cobra.Command{
 	Long: `The "rke2" command allows you to install, manage, and uninstall RKE2.
 
 Examples:
-  edgectl rke2 server install       # Install RKE2 Server
-  edgectl rke2 agent install        # Install RKE2 Agent
-  edgectl rke2 system purge         # Uninstall RKE2
+  edgectl rke2 server install        # Install RKE2 Server
+  edgectl rke2 agent install         # Install RKE2 Agent
+  edgectl rke2 system purge          # Uninstall RKE2
+  edgectl rke2 system kubeconfig     # Fetch kubeconfig from Vault
+  edgectl rke2 system bash           # Configure bash environment
 `,
 }
 
@@ -35,7 +36,6 @@ func init() {
 	// Add all the modularized command packages
 	rke2Cmd.AddCommand(servercmd.Cmd)
 	rke2Cmd.AddCommand(agentcmd.Cmd)
-	rke2Cmd.AddCommand(configcmd.Cmd)
 	rke2Cmd.AddCommand(systemcmd.Cmd)
 	rke2Cmd.AddCommand(lbcmd.Cmd)
 }
